fix: remove indentation on 'body' block
The 'body' contains preformatted code blocks, rendered like so:

```
   <div>{{start_code}}</div>
```

If `start_code` contains newlines (let's say it's
`"first_line\nsecond_line"`), the template will render as follows:

```
   <div>first_line
second_line</div>
```

Which is fine, but if the `indent` filter is used to reindent this
block, the final HTML changes to:

```
   <div>first_line
     second_line</div>
```

Which means the editor will see the following code:

```
first_line
      second_line
```

First of all this looks ugly, but in a whitespace-sensitive language
this is even a syntax error.

It is therefore not safe to reindent generated HTML code willy-nilly.

If you want to pretty-print HTML, you should use JavaScript to inject
the startup code into the editor (as the pretty-printer will not touch
a string embedded in JavaScript).