chore: front application with nginx
Front the Flask application `nginx` to do the experimental
proxying to different backends. Based on the percentage, the
traffic is either routed to the "current" server or the "test"
server.

Also switch Python version to 3.7.10

This is required for the newer `heroku-20` dyno (which hedy-test
uses) and also works on the older `heroku-18` dynos (which the other
hedy instances use.