feat: palette buttons append code instead of replacing
Right now, clicking a palette button complete replaces whatever
code is in the editor, destroying whatever they might have typed
already and losing their work.

This change will **append** the example code to whatever is in the code
editor, so the user can get the benefit of trying out the code without
destroying what they already have.  Probably goes well together with
PR #315, which really turns the palette into a palette of options to mix
and match.

I can also see a downside: this feature makes it possible to finish
exercises by clicking the right button+lightly editing the example.
Kids might use this to skip having to type out the entire syntax of
the command, thereby not getting the full learning experience.