chore: proper version control of Tailwind
It was a mistake of me to not insist on proper version control
of Tailwind from the start.

The result was that the colors subtly changed between v1 and v2 of
Tailwind, and the whole website would be styled to the colors of
the version of Tailwind that the last person to run `generate-css`
happened to have on their machine.

That's not great!

Instead, now do proper version control by using `package.json`,
and control the version of Tailwind we use.

For now, we settle on Tailwind v1, because it has the colors we
want. We could eventually decide to move to Tailwind v2, if we find
a way to restore the old colors.

For now, there doesn't seem to be a pressing reason to upgrade, so
this is the simplest way to get the result we want. We can always
consciously decide to upgrade later.