add central security settings for cookies
We aren't currenty setting samesite values for cookies, and that would be a good thing to do (https://flask.palletsprojects.com/en/1.1.x/security/).

This commit sets the cookie values for samesite, httponly, and secure in a central place so we don't need to worry about it if we add more cookie values.