feat: add 'query logging', structured logging per request
This introduces structured request logging, with a way to log
counters, arbitrary values, and timers within each request.

Scopes are automatically opened and closed matching Flask requests,
by using `@app.before_request` and `@app.teardown_request`.

Currently annotates a number of functions such as `db_get()`,
`render_template()`, etc etc with `@querylog.timed` which automatically
records the time taken inside those functions, and how often they are
called. Some counters and values are recorded manually with
`log_counter` and `log_value`.

I'd recommend everyone log as many things as they can that may impact
performance or stability of the app, to help with ease of debugging
later on.

The mechanism to add a log sink are in place, but I'm not sure what
the current preferred log target is (I see something has been done
involving Papertrail but I'm not sure what that is or how that works),
and I investigated logging these to JSONBin but we may run into issues
there (it's not strictly optimized for that).

S3 is probably an easy and cheap log target, but I'm open to other
suggestions.