chore: add `pylint` validation
Add `pylint` into the standard set of tests. `pylint` is a style
checker for Python. There are currently many style issues with the code
base, so the suite of generic style checking has been disabled. Instead,
we just check for obvious errors (misspellings of variables, etc) and
throw an error if we find any.

`pylint` found 2 in the current code base:

- Two functions in a testing class that had the same name (so the
  2nd function definition would overwrite the first). Fixed by changing
  the name of the 2nd function.
- Illegal access of `.line`, `.column`, `.allowed` on an instance of
  something that's declared as `Exception`. Some investigation teaches
  that the only Lark exception that has all 3 of these attributes is
  `UnexpectedCharacters`, so changed to catching only that exception.

Added scripts to do all of the validations we check:

```
build-tools/validate-python
build-tools/validate-tests
build-tools/validate-yaml
```

And a single script:

```
build-tools/validate
```

To run all validations.

All validations are now run (individually) in the PR validation workflow
script as well.