fix: save and restore querylog across restarts
Sometimes our request handling takes more than 30 seconds, at
which point in time `gunicorn` shoots our poor worker process in the
head. This means the logs never get written and we never get to figure
out what was taking so long.

When this happens, use a gunicorn shutdown hook to save pending log
records and flush the queue to disk, so the next run of the program
can try to submit it.

Also in this change, recording some more values that might come in
useful when debugging (performance) issues:

- Record the PID in the records
- Record the load average in the records