chore: run e2e tests during build
Make it so the end-to-end tests can be run during build.

We have to abstract away from the Cloud infrastructure necessary to run
all the tests (specifically: DynamoDB), and replace it with an in-memory
database.

Not entirely sold on the current approach yet, but at least it works
*somewhat*.

Not completely done, the code is now trying to send an email.