fix: add caching on static content (#527)
About 30% of the total request volume to the server is to retrieve
static error messages to use in the JavaScript (`/error_messages.js` and
`/auth/texts`). Because they are not cached, they are currently loaded
upon every page load (every time a user hits **Next level**).

They could *easily* be cached though, bringing down the total request
volume to the server by a lot.

Add a caching header to make the browser cache these responses for an
hour; we could think about routing them through the CDN as well,
but that's perhaps a bit much for now. Let's just see how this performs.