fix(heroku): new `db_init` method not called on Heroku
On Heroku, we don't use the `if __name__ == '__main__':` block to
initialize our application, but instead run the server through
`gunicorn`.

The new `db_init()` function introduced in #537, that is required to be
called at the start of the process, was not called when loaded via
gunicorn.

Add a new hook to properly call initialization routines in both
scenarios.