fix: script errors aren't reported properly in prod
Because `app.js` is loaded from the CDN in prod, if there's an error in
the script the browser hides the details of the error for security
reasons (different domain = you don't get to see the details).

Add a cross-origin header so the browser will treat the external
resources as safe.