chore: try to get browser to report Script Errors, part 2
Client script errors are still reported as "Script Error",
indicating that the browser does not trust the source of the
script.

We added `access-control-allow-origin: *` already, but this
is not sufficient. Some other random webpage I found says we also
have to add `crossorigin="anonymous"`, so let's give that a shot.