fix: level 3 syntax highlighting (#579)
The regexes in the tokenizer will only ever see one line at a time,
without the closing `\n`. As soon as all characters in a line have
been consumed, the tokenizer moves to the next line, remembering
the current state.

This means that consuming 0 characters at the end of line to move
back to the start state is not possible.

Instead, we take advantage of the ability of the tokenizer to capture
multiple groups (characters between `(...)` in the regex) and assign all
of them token classes at once by specifying them in arrays. This allows
us to finish regexes with `(.*)` to consume anything till the end of the
line (if present), so we know that we can move back to the `start` state
after having done so.

Co-authored-by: Felienne <felienne@gmail.com>