feat: "try example" button on examples
For ever example we find in the content of tabs, add a button
to it to quickly copy the code into the editor.

Also in this commit: renamed `error_messages.js` to
`client_messages.js`, since sometimes we need UI translations in the
JavaScript as well, and it's a better name.