chore: remove 'grammars-Total' files
The files were ignored, but not actually removed yet, so actually
remove them.

Since git only tracks files and not directories, removing all files
would not create the directory on checkout, and writing the `Total`
files would fail because of a missing directory.

So put a single empty placeholder file (`.keepdir`) in the directory
so that git will create the dir.