fix: load_yaml_pickle test fails if a pickled file is on disk
The pickle test that loads `hu.yaml` compares it to loading
`hu.yaml.pickle`; if there was a `.pickle` file remaining on
disk, it may have different contents than the `.yaml` file.

This will not happen on the server since it gets a clean redeploy
every time, but it may happen on dev machines.