fix: can only concatenate str (not "ParseResult") to str
This is showing up in the logs as 500 errors.

Theoretical debugging session: this happens because `transpile_inner`
now returns a `ParseResult` instead of a string, while some places
are still expecting it to return the old string.

Only occurs occasionally because it only happens when we have a parsing
level fallback case happening.