Basic fix for 745 (#755)
This updates the print code to always use f-string to allow for lists to be printed. That makes the tests easier to read and prevents errors. However, in a near future I will to implement a basic type system so we can tell kids they should not print a whole list cause why would you even at this stage. But at least this prevents ugly Python errors.