chore: do Tailwind build at Heroku deploy
We do a minimized Tailwind build: this strips away CSS classes that we
don't use, and makes the difference between a CSS file of 40kB and 2MB (!).

However, it's annoying during development, because anytime you
experiment with using a new Tailwind class you have to remember to
re-generate the CSS files.

Instead, we'll just commit the full Tailwind source file of 2MB into
development, and do the optimized build automatically during Heroku
deployent. That way, the CSS is still efficient in production but the
developer experience is better.