chore: re-enable all tailwind plugins
Originally, I had manually disabled a number of Tailwind plugins
in the config, in an attempt to keep the size of the resulting CSS
file down.

Later on, we moved to the better solution of stripping unused CSS
classes during the build, but I had not turned the plugins back on.

The result is that we *never* emitted "opacity" classes, and so they
could not be used.

This PR re-enables everything, CSS stripping is the correct and better
solution to controlling the size of the CSS bundle.

Also in this PR: change the positioning of the modal content to `fixed`
instead of `absolute`, so you can't scroll the modal away.