Improved 'else' syntax highlighting to be level-dependent (#1006)
In the original fix we replaced the space required of on the 'else' statement. However, on levels 4 till 7 the corresponding code to the else is on the same line as the command, so a space is required. In this PR the 'syntaxModesRules.js' rules are adjusted to provide for this change. Can be tested by using the else command on a level below 7 and on 7 and higher.