fix: signup fails if languages are selected
If any "previous experience" languages are selected, the signup button
doesn't do anything.

There was a programming mistake around using `this` in JavaScript:

```ts
$('input[name=languages]').filter(':checked').map(function() { return $(this).val() as string }).get()
```

Would have been correct, because `function()` doesn't bind `this` (and
so `this` would have referred to the checkbox.

However, we wrote this:

```ts
$('input[name=languages]').filter(':checked').map(() => $(this).val() as string).get()
```

And because a `() => { }` arrow function DOES bind `this` to whatever
it's currently bound to, we wouldn't be bound to the checkbox but to the
global object.

Instead, just make the actual argument explicit, and don't rely on
`this` at all. It's always tricky in JS!