fix: higlight error location in line
This is a resubmission of #1147, which had done most of the heavy lifting
to figure out what needed to happen already but had some unnecessary
changes in there as well.

The `Location` field in the error response already contained both row
and column number, so good enough to achieve column-specific
highlighting.

To support `InvalidCommandException`, which *can* only have a row and
not a column (since we don't have that information available), we
introduce the convention that `Location` can be either `[row, col]` or
just `[row]`. The client will highlight the entire line or just a single
character.

(I added a helper future-proofed a bit by adding a parameter so that
once we want to highlight whole words, we can do that too).

There is a bit of a short-term hack in `InvalidCommandException`,
which should be tackled as part of addressing #1178.