introduce js-validated-form class
In pull-request #1183, the code that populates the `UserForm` was changed to lift the values from the elements of `form.auth` to `form#signup`. Unfortunedly this effected the the login page as well.

In this pull-request we have introduced the `js-validated-form` class. It is a marker class that signals

* It is used by JavaScript, not only for styling purposes
* It concerns validated forms

We can now revert to a class-based selection of forms.