fix: change quiz database key schema
Change the keys schema for the quiz database. The new schema will
make it easier to query for the answers given by a single user.

The old schema was: `quizAttemptId -> {...answer data...}`, but since
attempt IDs were allocated as GUIDs, there was no meaning and
relationship to them (and hence: no possibility to query).

The new key schema is: `user -> levelAttempt -> {...answer data...}`.

This makes it trivial to query for all answers by a single user...
and if we ever support prefix queries on the range key it will also
be trivial to get the answers for a single level.