fix: quiz is broken if user is not logged in
If the user is not logged in, the username is `''`.

First of all, DynamoDB rejects this value with an error.

Second of all, if we would record all anonymous users under the same
key, the data partition would become too big.

Instead, we'll record the responses under a username like
`'anonymous:e91f5ac2a6c946e893683c5c02781117'`.