User information no longer necessary in `render_template` calls (#1188)
This pull-request encompasses the following improvements:

* It places the user on the Flask session.
* It automatically injects user information in the context for templates.
* It refactors `is_teacher` and `is_admin` to accept a user instead of a request.

Below you can find a detailed motivation for these changes.

This pull-request uses a [_context processor_](https://flask.palletsprojects.com/en/2.0.x/templating/#context-processors) to inject user information, such as `username`, `is_teacher` and `is_admin` into the context automatically. This relieves developers from passing along this information with each call to `render_template`.

The context processor reads the user information from the user on the [Flask session](https://flask.palletsprojects.com/en/2.0.x/api/#flask.session). The user is placed on the session when they login and is removed when they user logs out or when they are destroyed.

The `is_teacher` and `is_admin` function accepted a request. This request was used to retrieve the current user from the database. This is an example of a function with side-effects, which should be avoided. Therefore the functions are changed to accept a user instead.

This pull-request is related to issue #1179. It removes a few of the globals that hamper progress on that front.