chore: log query string in querylog
We used to be only logging the path in the query log (`/hedy/8`),
but the language is pretty relevant and that is passed in the query
string (`/hedy/8?lang=fr`).

Log the query string as well if passed.

Also add an example query to the download script so I don't have to
look up how to do this every time :).