chore: remove AWS god keys (#1327)
Currently, we are using service-specific AWS user keys. What does that
mean?

There is one AWS user who can has full admin access to DynamoDB, and
this user is used by both the alpha as well as the production
environment.

There is a separate AWS user who has full admin access to SES, and this
user is also used by both the alpha as well as the production
environment.

There are multiple things wrong with this setup:

- There are different keys per service we want to use, which is
  unnecessarily annoying to configure.
- The keys have unnecessarily broad permissions.
- The keys are not scoped to a single (alpha or prod) environment,
  and importantly scoped to the resources that that environment
  needs to access.

I've rectified the configuration in AWS itself. The current PR switches
over to just using the implicit, environment-provided keys that are tied
to the alpha and prod environments.

After this PR is merged, we can get rid of the god users and their keys.