refactor: remove 'menu' parameter
All calls to `render_template()` used to include a `menu=` parameter,
to render the entries of the main menu (in every language).

It was an annoying requirement to pass in the menu everywhere, and
unnecessary because the menu could just be a global variable in the
template.

Add a template global `main_menu_entries()` which returns the same
as what `render_main_menu()` used to do. Instead, `menu` is now
a boolean that controls whether or not the menu should be rendered
at all, and the default is `True` so in most cases you don't need
to pass anything: you only need to pass `menu=False` in cases where
you don't want the menu rendered, which is only for embedded frames
like for the quiz.

Relates to #1179.