Whats next update NL and EN (#1385)
* Update whats next NL

Added more levels for what's next and updatet the previous levels of the what's next adventure in Dutch.

* Update what's next EN

I've updatet the what's next adventures so they no longer say 'in level 3' but 'in this level' or 'in the next level'. So they are change-proof.

I have not filled in level 11 to  17 yet, so they're empty.  I will do that soon.

* fix small issue and stop checking next adventures

Co-authored-by: Felienne <felienne@gmail.com>