fix: automatically log users in from long-lived cookies (#1457)
The Hedy code base was operating in a split-brain scenario:

* Some code would inspect the session to find the currently logged
  in user. Sessions get cleared when the browser is closed.
* Some code would look at the long-lived login cookie and perform
  database lookups on every request to figure out what the currently
  logged-in user is.

The user was purposely stored in the session to save database calls
on every request, but if you close and reopen the browser the two
code paths would disagree:

- According to the session, you are now logged out.
- According to the login cookie, you are still logged in.

This PR unifies the behavior. The new behavior is as follows:

- All code will look at the session to determine whether you are logged
  in or not.
- The session would be cleared when you close the browser. In order to
  support persistent login sessions, if we detect you *don't* have
  a session user but you *do* have a login cookie, we immediately log
  you in from the cookie.

This restores the old behavior, with the advantage that we still save a
lot of database calls by keeping login information in the session.

Also contains a fix for development mode: `SECRET_KEY` will be set
to a consistent value automatically, so that if you restart your local
dev server your session doesn't automatically invalidate.

Fixes #1425.