chore: remove unused translations
During translation cleanup, I came across a number of translations
that are no longer being used.

It seems unfair to ask future translators to translate these unused
sentences, so I'm removing them.

I suppose something might be said for removing it from `en` and not
the rest, but we still have git history if we ever want these back
anyway.