Adds types to level 10 and 11 in SEMANTICS (#1707)
* natural language semantics, level 2 en 3 added, level 1 little correction

* natural language semantics, fixed little mistake in table layout

* level 2 split into two, so we have semantics descriptions up to level 4

* add types to semantics description for level 1 and 2

* corrections to types in level 1 and 2 in semantics description

* add types to level 3 and 4 in semantics description

* tiny change in type for command is

* adds types to level 5 and 6

* adds types to level 7, 8, 9 in SEMANTICS description

* Add empty type

For level 1 t/m 9, adds empty as type used with commands forward and turn, also for echo in level 1. Also deletes the possibility of an incomplete exception for the echo command in level 1.

* Adds types to level 10 and 11

The new command is `for`. In level 10, this can only be used as `for i in list`, but in level 11 it can also be used as `for i in range j to k`. The others did not change, but in level 11, forward and turn disappear.

Co-authored-by: Felienne <felienne@gmail.com>