fix: JavaScript broken on Safari and old Firefox
The lookbehind regex we were using to match a word boundary in non-ASCII
languages is only supported in Chrome, not in other browsers.

Switch to an alternative that isn't as correct (because it must consume
the non-word character instead of just assert it), but should do fine
in most cases.