[BUG] Support numeric values in list assignment
In order to support lists that fully consist of numbers, this commit
will not surround list values with quotes if all the values are in
a numerical format.

This will allow Hedy programs to use values from lists that require
the value to be a number, such as `forward` in turtle programs.