[BUG] Handle non-numerical values for repeat
This commit will introduce a new step to using repeats; we will assign
the amount of times we are repeating the statements to a temporary
variable. First, we try to convert the 'times' value to a number. If
this fails, we set it to 0. Effectively communicating; "we can only
count numbers".