Only show language switcher on ace when admin (#3327)
**Description**
This PR removes the languages switcher on the ace editor for normal users. It has become a bit redundant, especially with the already existing keyword language switcher on the dropdown bar. We also remove the 'hide keyword switcher' option from the class customizations.

**Fixes**
This PR fixes #3322

**How to test**
Verify that the language switcher is still shown when being the admin, otherwise it should be hidden.