[FEATURE] Allows teachers to hide/show keyword switcher in menu bar (#3355)
**Description**
We partly re-implement the code deleted in #3327 where we decided to remove the language switcher from the code editor. We now re-introduce a "hide keyword switcher" but instead aim on the menubar dropdown keyword switcher. The back-end code can remain the same, we only had to make some adjustments to the front-end.

**Fixes _issue or discussion number_**
This PR fixes #3354.

**How to test**
Make sure you are in a class and have a keyword-supported-non-english profile language. Verify that you can see and use the keyword switcher dropdown from the menubar. When hidden using the class customizations the button should been gone.