Add new hedy paper by Gilsing et al (#3364)
First, provide a one line summary of your changes in the **Title of the PR**

**Description**

This PR adds the new paper about Hedy: https://www.sciencedirect.com/science/article/pii/S2590118422000557

**Fixes #3324**

**How to test**

N/A

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [x] Describes changes in the format above (present tense)
- [x] Links to an existing issue or discussion 
- [x] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!