Fixes translation of keywords with several  options (#3371)
First, provide a one line summary of your changes in the **Title of the PR**

**Now fill out the remainder of this template by replacing all _italic_ content**

**Description**

_Changes in detail_

Fixes the translation of keywords that have several options, for example repeat in French can be translated as "répète" and "repete".  This was done by changing the dictionary in the translator from a dictionary that maps strings to strings, to one that maps strings to list of strings. There's one thing I don't feel particularly happy about: while we can translate repete back and forth, we won't get the same result, it means that we can't use repete in multi-level testing because when it's translated from English back to French it will return "répète", meaning that the test witll fail because technically is not the same string. 

**Fixes #2711 **

**How to test**

I added a few tests

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [x] Describes changes in the format above (present tense)
- [x] Links to an existing issue or discussion 
- [x] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!