Improves readability on mobile devices (#3380)
Makes sure fonts are displayed properly and also uses full width on smaller devices for adventure texts and example codes. Also keeps the logo nice and round.

**Description**

Makes sure fonts are displayed properly and also uses full width on smaller devices for adventure texts and example codes. Also keeps the logo nice and round.

There's probably a lot more that can be done, but this should be considered a first step.

**Fixes #3379**

Always link the number of the issue or of the discussion that your pr concerns.

**How to test**

* Use the inspector in your browser and enable mobile view (how to do that depends on the browser). Also you could use a mobile device to view the Hedy version on your computer. For that they need to be on the same network.

The changes are template only, so it should be easy to compare current and new situations.

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [x] Describes changes in the format above (present tense)
- [x] Links to an existing issue or discussion 
- [x] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!