[FEATURE] Adds option for teacher to set quiz threshold for unlocking next level (#3362)
**Description**
In this PR we add an exciting new feature: quiz thresholds! Teachers are able to set a quiz score between 0 and 100 that is the minimal score needed for the user to "unlock" the next level. When the score is not reached yet the level is unlocked and the 'next level' button is hidden. This should make it easier for teachers to make multiple levels available but still ensure that children have an understanding of a level before skipping to the next one.

**Fixes**
This PR fixes #3298

**How to test**
Make sure you are a teacher and navigate to the customization of a class. Notice the new "unlock thresholds" option on which you can set a quiz value between 0 and 100. Set it and verify with a student account that this indeed works as expected. When trying to directly navigating to a non-locked and not-achieved level the following error should be shown:

`Quiz threshold not reached to unlock this level`