Allows sending emails in other languages (#3393)
**Description**

We usually use the language of the current user when sending emails, but that is not suitable in all cases. Specifically if an admin needs to email a user, this should be done in the language of the user.

This fixes the issue when an admin marks a user as teacher, and also when an admin changes the email address of a user.

**Fixes #3289 **

Always link the number of the issue or of the discussion that your pr concerns.

**How to test**

As stated in the issue, this is hard to test locally because sending emails does not work locally. This would need to be tested on alpha.

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [x] Describes changes in the format above (present tense)
- [x] Links to an existing issue or discussion 
- [ ] Has a "How to test" section