[TESTING] Excludes code tests when no changes are made there (#3456)
We improve our tests by excluding the code snippet tests when there are no relevant changes made in either the /content or the /grammars folder. This reduces our test time significantly, making it easier to work with and reduce the waiting time for validation when working on the front-end / unrelated back-end.