[BUGFIX] Dissabling autocomplete for changing password (#3426)
**Description**

Adds the attribute `autocomplete="new-password"` to the input fields in my-account page to avoid then autocompleting the password.

**Fixes #3283**

Always link the number of the issue or of the discussion that your pr concerns.

**How to test**

* Go to http://0.0.0.0:8080/my-profile and try changing your password

**Checklist**
Done? Check if you have it all in place using this list:*
  
- [ ] Describes changes in the format above (present tense)
- [ ] Links to an existing issue or discussion 
- [ ] Has a "How to test" section

If you're unsure about any of these, don't hesitate to ask. We're here to help!